+++ src/components/Table/EditableTable/types.tsx
@@ -1,0 +1,7 @@
+export interface Stocks {
+  symbol: string;
+  buyRate: number;
+  kitta: number;
+  ltp: number;
+  total: number;
+}
